requirements.txt
anthropic
python-dotenv
markdown
